+++ CHANGELOG.md
@@ -63,6 +63,15 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 * [PR-424](https://github.com/itk-dev/hoeringsportal/pull/424)
   2945: Cleaned up form templates. Applied security updates.
 
+## [4.7.0] - 2025-03-10
+
+* [PR-455](https://github.com/itk-dev/hoeringsportal/pull/455)
+  * Fixed bug in display of "on behalf of” on hearing reply details
+  * Added "on behalf of” on hearing reply list
+  * Corrected handling of ticket creation time
+* [PR-454](https://github.com/itk-dev/hoeringsportal/pull/454)
+  Update person name in Deskpro on email match
+
 ## [4.6.5] - 2025-03-05
 
 * [PR-456](https://github.com/itk-dev/hoeringsportal/pull/456)
@@ -338,7 +347,8 @@ Updated drupal core 8.6.16
 
 Initial release
 
-[Unreleased]: https://github.com/itk-dev/hoeringsportal/compare/4.6.5...HEAD
+[Unreleased]: https://github.com/itk-dev/hoeringsportal/compare/4.7.0...HEAD
+[4.7.0]: https://github.com/itk-dev/hoeringsportal/compare/4.6.5...4.7.0
 [4.6.5]: https://github.com/itk-dev/hoeringsportal/compare/4.6.4...4.6.5
 [4.6.4]: https://github.com/itk-dev/hoeringsportal/compare/4.6.3...4.6.4
 [4.6.3]: https://github.com/itk-dev/hoeringsportal/compare/4.6.2...4.6.3